nlu_flow/sentence_classification/slang_classifier/ludwig_fine_tune/requirements.txt
fastapi
uvicorn
requests
pandas
dill

setuptools>=41.0.0
numpy==1.18.5
tensorflow-gpu
ludwig[serve,text]


